% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_icils.R
\name{download_icils}
\alias{download_icils}
\title{Download and import an empty FDZ file.}
\usage{
download_icils(
  year = c("2018", "2013"),
  data_type = c("stud_dat", "stud_nat_dat", "stud_int_dat", "teach_dat", "teach_nat_dat",
    "teach_int_dat", "school_dat", "school_nat_dat", "school_int_dat", "it_dat")
)
}
\arguments{
\item{year}{Year of the assessment.}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty ICILS data set from the \href{https://www.iqb.hu-berlin.de/fdz/studies/}{FDZ homepage}.
}
\details{
The function downloads and imports an empty ICILS data set (\code{Leerdatensatz}) from the FDZ homepage.
These data sets contain zero rows.
The data is imported via \code{\link[eatGADS]{import_spss}} as a \code{GADSdat} object.
}
\examples{
fdz_icils <- download_icils(year = "2018",
                        data_type = "stud_dat")
}
