% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_iglu.R
\name{download_iglu}
\alias{download_iglu}
\title{Download and import an empty FDZ file.}
\usage{
download_iglu(
  year = c("2016", "2011", "2006", "2001"),
  data_type = c("stud_par_dat", "teach_ger_dat", "teach_math_dat", "teach_gen_dat",
    "teach_dat", "teach_stud_dat", "school_dat", "tracking", "testscores")
)
}
\arguments{
\item{year}{Year of the assessment.}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty IGLU data set from the \href{https://www.iqb.hu-berlin.de/fdz/studies/}{FDZ homepage}.
}
\details{
The function downloads and imports an empty IGLU data set (\code{Leerdatensatz}) from the FDZ homepage.
These data sets contain zero rows.
The data is imported via \code{\link[eatGADS]{import_spss}} as a \code{GADSdat} object.
}
\examples{
fdz_iglu <- download_iglu(year = "2016",
                        data_type = "stud_par_dat")
}
