% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pisa.R
\name{download_pisa}
\alias{download_pisa}
\title{Download and import an empty FDZ file.}
\usage{
download_pisa(
  year = c("2018", "2015", "2012", "2009", "2006", "2003", "2000"),
  data_type = c("stud_dat_9kl", "stud_dat_15j", "stud_par_dat", "stud_par_dat_9kl",
    "stud_par_dat_15j", "par_dat", "par_dat_9kl", "par_dat_15j", "teach_dat",
    "teach_dat_9kl", "teach_dat_15j", "school_dat", "school_dat_9kl", "school_dat_15j",
    "matching", "timing")
)
}
\arguments{
\item{year}{Year of the assessment.}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty PISA data set from the \href{https://www.iqb.hu-berlin.de/fdz/studies/}{FDZ homepage}.
}
\details{
The function downloads and imports an empty PISA data set (\code{Leerdatensatz}) from the FDZ homepage.
These data sets contain zero rows.
The data is imported via \code{\link[eatGADS]{import_spss}} as a \code{GADSdat} object.
}
\examples{
fdz_pisa <- download_pisa(year = "2018",
                        data_type = "stud_par_dat_9kl")
}
