% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_timss.R
\name{download_timss}
\alias{download_timss}
\title{Download and import an empty FDZ file.}
\usage{
download_timss(
  year = c("2019", "2015", "2011", "2007"),
  data_type = c("stud_par_dat", "teach_dat", "teach_stud_dat", "school_dat", "tracking")
)
}
\arguments{
\item{year}{Year of the assessment.}

\item{data_type}{Type of the data.}
}
\description{
Download and import an empty TIMSS data set from the \href{https://www.iqb.hu-berlin.de/fdz/studies/}{FDZ homepage}.
}
\details{
The function downloads and imports an empty TIMSS data set (\code{Leerdatensatz}) from the FDZ homepage.
These data sets contain zero rows.
The data is imported via \code{\link[eatGADS]{import_spss}} as a \code{GADSdat} object.
}
\examples{
fdz_pisa <- download_timss(year = "2019",
                        data_type = "stud_par_dat_9kl")
}
